chore: Simplify SVG validation
