+++ packages/snaps-sdk/src/internals/svg.test.ts
@@ -1,30 +1,12 @@
-import { parseSvg } from './svg';
+import { is } from 'superstruct';
 
-const SNAP_ICON =
-  '<svg width="24" height="25" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M17.037 0H6.975C2.605 0 0 2.617 0 6.987v10.05c0 4.37 2.605 6.975 6.975 6.975h10.05c4.37 0 6.975-2.605 6.975-6.976V6.988C24.012 2.617 21.407 0 17.037 0ZM11.49 17.757c0 .36-.18.684-.492.876a.975.975 0 0 1-.54.156 1.11 1.11 0 0 1-.469-.108l-4.202-2.1a1.811 1.811 0 0 1-.985-1.61v-3.973c0-.36.18-.685.493-.877a1.04 1.04 0 0 1 1.008-.048l4.202 2.101a1.8 1.8 0 0 1 .997 1.609v3.974h-.012Zm-.252-6.423L6.723 8.896a1.045 1.045 0 0 1-.528-.924c0-.384.204-.744.528-.924l4.515-2.438a1.631 1.631 0 0 1 1.524 0l4.515 2.438c.324.18.528.528.528.924s-.204.744-.528.924l-4.515 2.438c-.24.132-.504.192-.768.192a1.54 1.54 0 0 1-.756-.192Zm7.972 3.638c0 .684-.385 1.308-.997 1.608l-4.202 2.101c-.144.072-.3.108-.468.108a.975.975 0 0 1-.54-.156 1.017 1.017 0 0 1-.493-.876v-3.974c0-.684.384-1.309.997-1.609l4.202-2.101a1.04 1.04 0 0 1 1.008.048c.313.192.493.516.493.877v3.974Z" fill="#24272A"/></svg>';
+import { svg } from './svg';
 
-describe('parseSvg', () => {
-  it('parses valid SVGs', () => {
-    expect(parseSvg(SNAP_ICON)).toMatchInlineSnapshot(`
-        {
-          "@_fill": "none",
-          "@_height": 25,
-          "@_width": 24,
-          "@_xmlns": "http://www.w3.org/2000/svg",
-          "path": {
-            "@_d": "M17.037 0H6.975C2.605 0 0 2.617 0 6.987v10.05c0 4.37 2.605 6.975 6.975 6.975h10.05c4.37 0 6.975-2.605 6.975-6.976V6.988C24.012 2.617 21.407 0 17.037 0ZM11.49 17.757c0 .36-.18.684-.492.876a.975.975 0 0 1-.54.156 1.11 1.11 0 0 1-.469-.108l-4.202-2.1a1.811 1.811 0 0 1-.985-1.61v-3.973c0-.36.18-.685.493-.877a1.04 1.04 0 0 1 1.008-.048l4.202 2.101a1.8 1.8 0 0 1 .997 1.609v3.974h-.012Zm-.252-6.423L6.723 8.896a1.045 1.045 0 0 1-.528-.924c0-.384.204-.744.528-.924l4.515-2.438a1.631 1.631 0 0 1 1.524 0l4.515 2.438c.324.18.528.528.528.924s-.204.744-.528.924l-4.515 2.438c-.24.132-.504.192-.768.192a1.54 1.54 0 0 1-.756-.192Zm7.972 3.638c0 .684-.385 1.308-.997 1.608l-4.202 2.101c-.144.072-.3.108-.468.108a.975.975 0 0 1-.54-.156 1.017 1.017 0 0 1-.493-.876v-3.974c0-.684.384-1.309.997-1.609l4.202-2.101a1.04 1.04 0 0 1 1.008.048c.313.192.493.516.493.877v3.974Z",
-            "@_fill": "#24272A",
-          },
-        }
-      `);
-  });
-
-  it('returns an empty object for empty SVGs', () => {
-    expect(parseSvg('<svg />')).toStrictEqual({});
-  });
-
-  it('throws for invalid SVGs', () => {
-    expect(() => parseSvg('')).toThrow('Snap icon must be a valid SVG.');
-    expect(() => parseSvg('foo')).toThrow('Snap icon must be a valid SVG.');
+describe('svg', () => {
+  it('validates an SVG string', () => {
+    expect(is('<svg />', svg())).toBe(true);
+    expect(is('<svg></svg>', svg())).toBe(true);
+    expect(is('<foo/>', svg())).toBe(false);
+    expect(is(1, svg())).toBe(false);
   });
 });
