+++ packages/snaps-utils/coverage.json
@@ -1,6 +1,6 @@
 {
-  "branches": 96.73,
-  "functions": 98.75,
-  "lines": 98.83,
-  "statements": 94.85
+  "branches": 96.74,
+  "functions": 98.76,
+  "lines": 98.84,
+  "statements": 94.92
 }
