+++ packages/snaps-utils/package.json
@@ -66,6 +66,7 @@
     "cron-parser": "^4.5.0",
     "fast-deep-equal": "^3.1.3",
     "fast-json-stable-stringify": "^2.1.0",
+    "fast-xml-parser": "^4.3.4",
     "marked": "^12.0.1",
     "rfdc": "^1.3.0",
     "semver": "^7.5.4",
